hello/FundooLoginPage/requirements.txt
aioredis==1.3.1
alabaster==0.7.12
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.2
Automat==0.8.0
autopep8==1.4.4
Babel==2.8.0
boto3==1.11.7
botocore==1.14.7
Brlapi==0.6.6
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
CodeConvert==3.0.2
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
daphne==2.4.1
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2
django-admin==2.0.0
django-admin-interface==0.12.0
django-colorfield==0.1.15
django-cors-headers==3.2.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-excel-response2==2.0.8
django-filter==2.2.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-models-ext==1.1.9
django-nine==0.2.3
django-permissions==1.2.2
django-redis==4.11.0
django-redis-cache==2.1.0
django-rest-auth==0.9.5
django-rest-authtoken==1.2.4
django-rest-passwordreset==1.0.0
django-short-url==1.1.4
django-six==1.0.4
django-storages==1.8
django-taggit==1.2.0
django-templated-mail==1.1.1
django-viewsets==0.1.9
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
docutils==0.16
elasticsearch==6.4.0
elasticsearch-django==6.5
elasticsearch-dsl==6.4.0
furl==2.1.0
futures==3.1.1
get==2019.4.13
hiredis==1.0.1
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.4.0
incremental==17.5.0
Jinja2==2.10.3
jmespath==0.9.4
jwt==0.6.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.1.0
msgpack==0.6.2
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
orderedmultidict==1.0.1
packaging==20.0
pbr==5.4.4
pexpect==4.2.1
Pillow==5.1.0
pipenv==2018.11.26
pluggy==0.13.1
post==2019.4.13
protobuf==3.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
public==2019.4.13
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pyee==6.0.0
Pygments==2.5.2
pygobject==3.26.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymitter==0.2.3
PyNaCl==1.1.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyRFC3339==1.0
pytest==5.3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-decouple==3.3
python-dotenv==0.11.0
python-memcached==1.59
python-oauth2==1.1.1
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
query-string==2019.4.13
redis==3.3.11
reportlab==3.4.0
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
s3==3.0.0
s3transfer==0.3.1
screen==1.0.1
SecretStorage==2.3.1
service-identity==18.1.0
shortuuid==0.5.0
simplejson==3.13.2
six==1.14.0
snowballstemmer==2.0.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
system-service==0.3
systemd-python==234
testresources==2.0.1
TimeConvert==1.5.1
Twisted==19.10.0
txaio==18.8.1
tzlocal==2.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.3
validate-email==1.3
virtualenv==16.7.9
virtualenv-clone==0.5.3
wadllib==1.3.2
wcwidth==0.1.8
xkit==0.0.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==1.0.0
zope.interface==4.7.1
